Flytter resten av Ressurser fra Testprosjekt og renamer litt
