+++ Difi.Oppslagstjeneste.Klient.Resources/Certificate/Data/DifiTestCertificateClient.p12
