+++ Difi.Oppslagstjeneste.Klient.Resources/Difi.Oppslagstjeneste.Klient.Resources.csproj
@@ -53,6 +53,7 @@
     <Compile Include="..\SolutionItems\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="Certificate\CertificateResource.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Xml\XmlResource.cs" />
     <Compile Include="Xml\XmlUtility.cs" />
@@ -74,6 +75,9 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Xml\Data\Response\SoapFault.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Certificate\Data\DifiTestCertificateClient.p12" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
