+++ Difi.Oppslagstjeneste.Klient.Tester/Difi.Oppslagstjeneste.Klient.Tests.csproj
@@ -87,7 +87,7 @@
     <Compile Include="DtoConverterTests.cs" />
     <Compile Include="Fakes\FakeHttpClientHandler.cs" />
     <Compile Include="Fakes\FakeHttpClientHandlerResponse.cs" />
-    <Compile Include="Integrasjon\Integrasjonstester.cs" />
+    <Compile Include="Smoke\SmokeTests.cs" />
     <Compile Include="OppslagstjenesteHelperTests.cs" />
     <Compile Include="Properties\AssemblyInfo.log4net.cs" />
     <Compile Include="Svar\ResponseContainerTests.cs" />
@@ -96,7 +96,7 @@
     <Compile Include="Utilities\CompareObjects\IComparator.cs" />
     <Compile Include="Utilities\CompareObjects\IDifference.cs" />
     <Compile Include="Utilities\DomeneUtility.cs" />
-    <Compile Include="Envelope\OppslagstjenestevalidatorMedStubbetSjekkTimestamp.cs" />
+    <Compile Include="Envelope\OppslagstjenestevalidatorWithStubbedCheckTimestamp.cs" />
     <Compile Include="Envelope\OppslagstjenesteValidatorTests.cs" />
     <Compile Include="KlientkonfigurasjonTests.cs" />
     <Compile Include="MiljøTests.cs" />
@@ -104,10 +104,10 @@
     <Compile Include="OppslagstjenesteKonfigurasjonTests.cs" />
     <Compile Include="Svar\EndringerSvarTester.cs" />
     <Compile Include="Svar\PersonerSvarTester.cs" />
-    <Compile Include="Domene\PersonTester.cs" />
+    <Compile Include="Domene\PersonTests.cs" />
     <Compile Include="Svar\PrintSertifikatSvarTester.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Domene\SoapExceptionTester.cs" />
+    <Compile Include="Domene\SoapExceptionTests.cs" />
     <Compile Include="Utilities\MockUtility.cs" />
     <Compile Include="Utilities\XmlUtility.cs" />
   </ItemGroup>
@@ -136,7 +136,6 @@
     <None Include="packages.config" />
     <EmbeddedResource Include="Testdata\difi-enhetstester.cer" />
     <EmbeddedResource Include="Testdata\difi-enhetstester.p12" />
-    <EmbeddedResource Include="Ressurser\Sertifikater\DifiTestsertifikatKlient.p12" />
   </ItemGroup>
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
