+++ Difi.Oppslagstjeneste.Klient.Tester/Envelope/OppslagstjenestevalidatorWithStubbedCheckTimestamp.cs
@@ -5,15 +5,15 @@ using Difi.Oppslagstjeneste.Klient.Svar;
 
 namespace Difi.Oppslagstjeneste.Klient.Tests.Envelope
 {
-    internal class OppslagstjenestevalidatorMedStubbetSjekkTimestamp : Oppslagstjenestevalidator
+    internal class OppslagstjenestevalidatorWithStubbedCheckTimestamp : Oppslagstjenestevalidator
     {
-        public OppslagstjenestevalidatorMedStubbetSjekkTimestamp(XmlDocument sendtDokument, ResponseContainer responseContainer,
+        public OppslagstjenestevalidatorWithStubbedCheckTimestamp(XmlDocument sendtDokument, ResponseContainer responseContainer,
             OppslagstjenesteKonfigurasjon oppslagstjenesteKonfigurasjon)
             : base(sendtDokument, responseContainer, oppslagstjenesteKonfigurasjon)
         {
         }
 
-        protected override void SjekkTimestamp(TimeSpan timeSpan)
+        protected override void CheckTimestamp(TimeSpan timeSpan)
         {
             //Do nothing.
         }
